Added xHCI host controller, USB core server, signal improvements, and DMA pool
- USB3: full xHCI implementation (xhci-mem, xhci-ring, xhci-port, caps), hub enumeration, host driver, DMA pool, combphy, USB IPC
- New usb-core server for higher-level USB device management
- Signal: sigsend syscall, improved signal delivery and restore
- Kernel: memory cache improvements, list fixes, barrier/fence primitives
- Misc: C++ test server, GitHub CI instructions, build tweaks

+++ kernel/heartbeat.c
@@ -10,7 +10,7 @@ struct heartbeat_led {
 
 #if defined(__SPACEMIT_K1__)
 static int pinctrl_set(uint32_t pin, uint32_t sel) {
-    uint32_t reg = PINCTRL_BASE + (pin*4);
+    uint64_t reg = PINCTRL_BASE + (pin*4);
 //    DEBUG("[PIN_CTRL] sel 0x%X => 0x%X\n", sel, reg);
     putreg32(sel, PA2DA(reg));
 //    DEBUG("[PIN_CTRL] 0x%X = 0x%X\n", reg, getreg32(reg));
